python/dns/requirements.txt
Fabric3==1.13.1.post1
boto3==1.9.139
dyn==1.8.1
termcolor==1.1.0
pyOpenSSL==19.0.0
decorator==4.4.0